Move files Books and Categories to components folder and create new files.
