--- src/redux/categories/Categories.js
@@ -1,14 +1,0 @@
-import React from 'react';
-
-function Categories() {
-  return (
-    <>
-      <h2>Categories</h2>
-      <button type="button" className="input-submit">
-        Check Status
-      </button>
-    </>
-  );
-}
-
-export default Categories;
